Update OneOCR v2.0 client library
