+++ ClientLibrary/lib/src/main/java/com/microsoft/projectoxford/vision/VisionServiceRestClient.java
@@ -35,8 +35,8 @@ package com.microsoft.projectoxford.vision;
 import com.google.gson.Gson;
 import com.microsoft.projectoxford.vision.contract.AnalysisInDomainResult;
 import com.microsoft.projectoxford.vision.contract.AnalysisResult;
-import com.microsoft.projectoxford.vision.contract.HandwritingRecognitionOperationResult;
-import com.microsoft.projectoxford.vision.contract.HandwritingRecognitionOperation;
+import com.microsoft.projectoxford.vision.contract.TextRecognitionOperationResult;
+import com.microsoft.projectoxford.vision.contract.TextRecognitionOperation;
 import com.microsoft.projectoxford.vision.contract.Model;
 import com.microsoft.projectoxford.vision.contract.ModelResult;
 import com.microsoft.projectoxford.vision.contract.OCR;
@@ -53,7 +53,7 @@ import java.util.Map;
 
 public class VisionServiceRestClient implements VisionServiceClient {
     private static final String DEFAULT_REGION = "westus";
-    private static final String DEFAULT_API_ROOT = "https://%s.api.cognitive.microsoft.com/vision/v1.0";
+    private static final String DEFAULT_API_ROOT = "https://%s.api.cognitive.microsoft.com/vision/v2.0";
     private final String apiRoot;
     private final WebServiceRequest restCall;
     private Gson gson = new Gson();
@@ -222,38 +222,38 @@ public class VisionServiceRestClient implements VisionServiceClient {
     }
 
     @Override
-    public HandwritingRecognitionOperation createHandwritingRecognitionOperationAsync(String url) throws VisionServiceException {
+    public TextRecognitionOperation createTextRecognitionOperationAsync(String url, String mode) throws VisionServiceException {
         Map<String, Object> params = new HashMap<>();
-        String path = apiRoot + "/RecognizeText?handwriting=true";
+        String path = apiRoot + "/RecognizeText?mode=" + mode;
         String uri = WebServiceRequest.getUrl(path, params);
 
         params.put("url", url);
         String operationUrl = (String) this.restCall.request(uri, "POST", params, null, false);
-        HandwritingRecognitionOperation HandwrittenOCR = new HandwritingRecognitionOperation(operationUrl);
+        TextRecognitionOperation OneOCR = new TextRecognitionOperation(operationUrl);
 
-        return HandwrittenOCR;
+        return OneOCR;
     }
 
     @Override
-    public HandwritingRecognitionOperation createHandwritingRecognitionOperationAsync(InputStream stream) throws VisionServiceException, IOException {
+    public TextRecognitionOperation createTextRecognitionOperationAsync(InputStream stream, String mode) throws VisionServiceException, IOException {
         Map<String, Object> params = new HashMap<>();
-        String path = apiRoot + "/RecognizeText?handwriting=true";
+        String path = apiRoot + "/RecognizeText?mode=" + mode;
         String uri = WebServiceRequest.getUrl(path, params);
 
         byte[] data = IOUtils.toByteArray(stream);
         params.put("data", data);
         String operationUrl = (String) this.restCall.request(uri, "POST", params, "application/octet-stream", false);
-        HandwritingRecognitionOperation HandwrittenOCR = new HandwritingRecognitionOperation(operationUrl);
+        TextRecognitionOperation OneOCR = new TextRecognitionOperation(operationUrl);
 
-        return HandwrittenOCR;
+        return OneOCR;
     }
 
     @Override
-    public HandwritingRecognitionOperationResult getHandwritingRecognitionOperationResultAsync(String uri) throws VisionServiceException {
+    public TextRecognitionOperationResult getTextRecognitionOperationResultAsync(String uri) throws VisionServiceException {
         String json = (String) this.restCall.request(uri, "GET", null, null, false);
-        HandwritingRecognitionOperationResult HandwrittenOCR = this.gson.fromJson(json, HandwritingRecognitionOperationResult.class);
+        TextRecognitionOperationResult OneOCR = this.gson.fromJson(json, TextRecognitionOperationResult.class);
 
-        return HandwrittenOCR;
+        return OneOCR;
     }
 
     @Override
